fuse-stat support for numeric long buckets + bug fixes in AsgStrategy and stat provider

+++ fuse-axis2/fuse-epb/src/test/java/com/kayhut/fuse/epb/plan/statistics/ElasticStatProviderTest.java
@@ -129,7 +129,7 @@ public class ElasticStatProviderTest {
 
         ageStatistics.forEach(bucketInfo -> {
             //Since this is a term: Lower bound value == Upper bound value
-            assertTrue((Double) bucketInfo.getLowerBound() <= (Double) bucketInfo.getHigherBound());
+            assertTrue((Long) bucketInfo.getLowerBound() <= (Long) bucketInfo.getHigherBound());
         });
     }
